+++ @xen-orchestra/lite/src/App.vue
@@ -32,6 +32,9 @@
 </template>
 
 <script lang="ts" setup>
+import { useUiStore } from "@/stores/ui.store";
+import { useActiveElement, useMagicKeys, whenever } from "@vueuse/core";
+import { logicAnd } from "@vueuse/math";
 import { difference } from "lodash";
 import { computed, ref, watch, watchEffect } from "vue";
 import favicon from "@/assets/favicon.svg";
@@ -58,13 +61,28 @@ link.href = favicon;
 
 document.title = "XO Lite";
 
-if (window.localStorage?.getItem("colorMode") !== "light") {
-  document.documentElement.classList.add("dark");
-}
-
 const xenApiStore = useXenApiStore();
 const hostStore = useHostStore();
 useChartTheme();
+const uiStore = useUiStore();
+
+if (import.meta.env.DEV) {
+  const activeElement = useActiveElement();
+  const { D } = useMagicKeys();
+
+  const canToggleDarkMode = computed(() => {
+    if (activeElement.value == null) {
+      return true;
+    }
+
+    return !["INPUT", "TEXTAREA"].includes(activeElement.value.tagName);
+  });
+
+  whenever(
+    logicAnd(D, canToggleDarkMode),
+    () => (uiStore.colorMode = uiStore.colorMode === "dark" ? "light" : "dark")
+  );
+}
 
 watchEffect(() => {
   if (xenApiStore.isConnected) {