+++ @xen-orchestra/lite/src/components/infra/InfraVmItem.vue
@@ -12,7 +12,7 @@
       :icon="faDisplay"
       :route="{ name: 'vm.console', params: { uuid: vm.uuid } }"
     >
-      {{ vm.name_label || '(VM)' }}
+      {{ vm.name_label || "(VM)" }}
       <template #actions>
         <InfraAction>
           <PowerStateIcon :state="vm?.power_state" />