+++ @xen-orchestra/lite/src/stores/vm.store.ts
@@ -1,15 +1,14 @@
-import { defineStore } from "pinia";
-import { computed } from "vue";
 import { sortRecordsByNameLabel } from "@/libs/utils";
 import type { GRANULARITY } from "@/libs/xapi-stats";
 import type { XenApiVm } from "@/libs/xen-api";
 import { useHostStore } from "@/stores/host.store";
 import { createRecordContext } from "@/stores/index";
 import { useXenApiStore } from "@/stores/xen-api.store";
+import { defineStore } from "pinia";
+import { computed } from "vue";
 
 export const useVmStore = defineStore("vm", () => {
   const hostStore = useHostStore();
-  const xapiStats = useXenApiStore().getXapiStats();
   const baseVmContext = createRecordContext<XenApiVm>("VM", {
     filter: (vm) =>
       !vm.is_a_snapshot && !vm.is_a_template && !vm.is_control_domain,
@@ -42,7 +41,7 @@ export const useVmStore = defineStore("vm", () => {
       throw new Error(`VM ${id} is halted or host could not be found.`);
     }
 
-    return xapiStats._getAndUpdateStats({
+    return useXenApiStore().getXapiStats()._getAndUpdateStats({
       host,
       uuid: vm.uuid,
       granularity,